escreen-tms authentication

+++ WEB.SERVICES/Service/GenericService.cs
@@ -1,9 +1,10 @@
 ﻿using AutoMapper;
 using FluentValidation;
 using LanguageExt;
+using System.Net;
 using WEB.DOMAIN.Interface;
-using WEB.SERVICES.DTO;
 using WEB.SERVICES.IService;
+using WEB.UTILITY.Helper;
 using WEB.UTILITY.Logger;
 
 namespace WEB.SERVICES.Service
@@ -32,35 +33,38 @@ namespace WEB.SERVICES.Service
             _logger = logger;
         }
 
-        public virtual async Task<Either<string, TDto>> GetByIdAsync(Guid id, CancellationToken ct = default)
+        public virtual async Task<Either<ApiResponse<string>, ApiResponse<TDto>>> GetByIdAsync(Guid id, CancellationToken ct = default)
         {
             try
             {
                 var entity = await _repository.GetByIdAsync(id, ct);
-                return entity == null ? "Entity not found" : _mapper.Map<TDto>(entity);
+                return entity == null
+                    ? Prelude.Left(ApiResponse<string>.Fail(["Entity not found"], HttpStatusCode.NotFound))
+                    : Prelude.Right(ApiResponse<TDto>.Ok(_mapper.Map<TDto>(entity)));
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error fetching entity by id");
-                return $"Error fetching entity: {ex.Message}";
+                return Prelude.Left(ApiResponse<string>.Fail(["Internal Server Error"], HttpStatusCode.InternalServerError));
             }
         }
 
-        public virtual async Task<Either<string, IEnumerable<TDto>>> GetAllAsync(CancellationToken ct = default)
+        public virtual async Task<Either<ApiResponse<string>, ApiResponse<IEnumerable<TDto>>>> GetAllAsync(CancellationToken ct = default, params string[] includePaths)
         {
             try
             {
-                var entities = await _repository.GetAllAsync(ct);
-                return _mapper.Map<IEnumerable<TDto>>(entities).ToList();
+                var entities = await _repository.GetAllAsync(ct, true, includePaths
+                );
+                return Prelude.Right(ApiResponse<IEnumerable<TDto>>.Ok(_mapper.Map<IEnumerable<TDto>>(entities).ToList()));
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error fetching all entities");
-                return $"Error fetching all entities: {ex.Message}";
+                return Prelude.Left(ApiResponse<string>.Fail(["Internal Server Error"], HttpStatusCode.InternalServerError));
             }
         }
 
-        public virtual async Task<Either<string, Guid>> AddAsync(TDto dto, CancellationToken ct = default)
+        public virtual async Task<Either<ApiResponse<string>, ApiResponse<Guid>>> AddAsync(TDto dto, CancellationToken ct = default)
         {
             try
             {
@@ -68,7 +72,7 @@ namespace WEB.SERVICES.Service
                 if (!validate.IsValid)
                 {
                     var errors = validate.Errors.Select(e => new { e.PropertyName, e.ErrorMessage });
-                    return string.Join("; ", errors.Select(e => $"{e.ErrorMessage}"));
+                    return Prelude.Left(ApiResponse<string>.Fail(errors.Select(x => x.ErrorMessage).ToList(), HttpStatusCode.InternalServerError));
                 }
 
                 var entity = _mapper.Map<TEntity>(dto);
@@ -78,16 +82,16 @@ namespace WEB.SERVICES.Service
                 }, ct);
 
                 var idProp = typeof(TEntity).GetProperty("Id");
-                return idProp != null ? (Guid)idProp.GetValue(entity)! : Guid.Empty;
+                return Prelude.Right(ApiResponse<Guid>.Ok(idProp != null ? (Guid)idProp.GetValue(entity)! : Guid.Empty));
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error adding entity");
-                return $"Error adding entity: {ex.Message}";
+                return Prelude.Left(ApiResponse<string>.Fail(["Internal Server Error"], HttpStatusCode.InternalServerError));
             }
         }
 
-        public virtual async Task<Either<string, bool>> UpdateAsync(TDto dto, CancellationToken ct = default)
+        public virtual async Task<Either<ApiResponse<string>, ApiResponse<string>>> UpdateAsync(TDto dto, CancellationToken ct = default)
         {
             try
             {
@@ -96,29 +100,33 @@ namespace WEB.SERVICES.Service
                 {
                     _repository.Update(entity);
                 }, ct);
-                return true;
+                return Prelude.Right(ApiResponse<string>.Ok("Update Successfully"));
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error updating entity");
-                return $"Error updating entity: {ex.Message}";
+                return Prelude.Left(ApiResponse<string>.Fail(["Internal Server Error"], HttpStatusCode.InternalServerError));
             }
         }
 
-        public virtual async Task<Either<string, bool>> DeleteListAsync(IEnumerable<Guid> ids, CancellationToken ct = default)
+        public virtual async Task<Either<ApiResponse<string>, ApiResponse<string>>> DeleteListAsync(IEnumerable<Guid> ids, CancellationToken ct = default)
         {
             try
             {
+                if (ids == null || !ids.Any())
+                {
+                    return Prelude.Left(ApiResponse<string>.Fail(["No IDs provided"]));
+                }
                 await _unitOfWork.ExecuteAsync(async c =>
                 {
                     await _repository.DeleteRangeAsync(ids, c);
                 }, ct);
-                return true;
+                return Prelude.Right(ApiResponse<string>.Ok("Deleted Successfully"));
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error deleting entities");
-                return $"Error deleting entities: {ex.Message}";
+                return Prelude.Left(ApiResponse<string>.Fail(["Internal Server Error"], HttpStatusCode.InternalServerError));
             }
         }
     }
